+++ interpreter/function/errors/errors.go
@@ -41,6 +41,18 @@ func CannotConvertToString(name string, num int) error {
 	return New(name, "Argument %d cannot convert to string because the value is literal", num)
 }
 
+// the error given when you querystring.get(req.url, "%ff");
+// fastly error is "Unexpected byte 0x%02X in UTF-8 in short-string"
+func UnexpectedByteInShortString(name string, num int) error {
+	return New(name, "Argument %d, after pct-decoding, is invalid utf-8", num)
+}
+
+// the error given when you querystring.get(req.url, "%eh");
+// fastly error is "Invalid hex char %c (0x%02X) in %%xx escape"
+func InvalidHexChar(name string, num byte) error {
+	return New(name, "Argument %d has invalid pct-encode sequence", num)
+}
+
 // Testing related errors
 type TestingError struct {
 	// Token info will be injected on interpreter