+++ interpreter/variable/hit.go
@@ -1,11 +1,7 @@
 package variable
 
 import (
-	"io"
-	"strings"
 	"time"
-
-	"net/http"
 
 	"github.com/pkg/errors"
 	"github.com/ysugimoto/falco/interpreter/context"
@@ -80,7 +76,9 @@ func (v *HitScopeVariables) Get(s context.Scope, name string) (value.Value, erro
 		return &value.Float{Value: 0.4}, nil
 	}
 
-	if val := v.getFromRegex(name); val != nil {
+	if val, err := v.getFromRegex(name); err != nil {
+		return nil, err
+	} else if val != nil {
 		return val, nil
 	}
 
@@ -92,10 +90,10 @@ func (v *HitScopeVariables) Get(s context.Scope, name string) (value.Value, erro
 	return val, nil
 }
 
-func (v *HitScopeVariables) getFromRegex(name string) value.Value {
+func (v *HitScopeVariables) getFromRegex(name string) (value.Value, error) {
 	// HTTP request header matching
 	if match := objectHttpHeaderRegex.FindStringSubmatch(name); match != nil {
-		return getResponseHeaderValue(v.ctx.Object, match[1])
+		return getResponseHeaderValue(v.ctx.Object, match[1]), nil
 	}
 	return v.base.getFromRegex(name)
 }
@@ -108,10 +106,16 @@ func (v *HitScopeVariables) Set(s context.Scope, name, operator string, val valu
 		}
 		return nil
 	case OBJ_RESPONSE:
+		// If synthetic response has already been set, ignore this assignment.
+		if v.ctx.HasSyntheticResponse {
+			return nil
+		}
+
 		if err := doAssign(v.ctx.ObjectResponse, operator, val); err != nil {
 			return errors.WithStack(err)
 		}
-		v.ctx.Object.Body = io.NopCloser(strings.NewReader(v.ctx.ObjectResponse.Value))
+		// obj.response indicates response header line, meand status text in http.Response
+		v.ctx.Object.Status = v.ctx.ObjectResponse.Value
 		return nil
 	case OBJ_STATUS:
 		i := &value.Integer{Value: 0}
@@ -119,7 +123,6 @@ func (v *HitScopeVariables) Set(s context.Scope, name, operator string, val valu
 			return errors.WithStack(err)
 		}
 		v.ctx.Object.StatusCode = int(i.Value)
-		v.ctx.Object.Status = http.StatusText(int(i.Value))
 		return nil
 	case OBJ_TTL:
 		if err := doAssign(v.ctx.ObjectTTL, operator, val); err != nil {